// zipkin-lens/src/types/action-types.ts
/*
 * Copyright 2015-2020 The OpenZipkin Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
 * in compliance with the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software distributed under the License
 * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 * or implied. See the License for the specific language governing permissions and limitations under
 * the License.
 */
import { Action } from 'redux';
import Dependencies from './Dependencies';

export enum ActionTypes {
  LOAD_DEPENDENCIES_REQUEST = 'LOAD_DEPENDENCIES_REQUEST',
  LOAD_DEPENDENCIES_SUCCESS = 'LOAD_DEPENDENCIES_SUCCESS',
  LOAD_DEPENDENCIES_FAILURE = 'LOAD_DEPENDENCIES_FAILURE',
  CLEAR_DEPENDENCIES = 'CLEAR_DEPENDENCIES',
}

interface LoadDependenciesRequestAction extends Action {
  type: ActionTypes.LOAD_DEPENDENCIES_REQUEST;
}

interface LoadDependenciesSuccessAction extends Action {
  type: ActionTypes.LOAD_DEPENDENCIES_SUCCESS;
  payload: {
    dependencies: Dependencies;
  };
}

interface LoadDependenciesFailureAction extends Action {
  type: ActionTypes.LOAD_DEPENDENCIES_FAILURE;
  payload: {
    error: any;
  };
}

interface ClearDependenciesAction extends Action {
  type: ActionTypes.CLEAR_DEPENDENCIES;
}

export type DependenciesAction =
  | LoadDependenciesRequestAction
  | LoadDependenciesSuccessAction
  | LoadDependenciesFailureAction
  | ClearDependenciesAction;